refactor: DTO 내 Long 타입을 Integer로 일괄 수정

+++ src/main/java/org/scoula/domain/chatbot/dto/ChatErrorDto.java
@@ -14,8 +14,8 @@ import java.time.LocalDateTime;
 @Builder
 @ApiModel("쳇봇 에러 저장 dto")
 public class ChatErrorDto {
-    private Long id;
-    private Long userId;
+    private Integer id;
+    private Integer userId;
     private String errorMessage;
     private String errorType;     // GPT / DB / API 등
     private LocalDateTime createdAt;
